backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
httpx==0.28.1
python-dotenv==1.1.1
pydantic==2.11.9
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
sentence-transformers==5.1.1
qdrant-client==1.15.1
pymupdf==1.26.4
python-docx==1.2.0
duckduckgo-search==8.1.1
groq==0.33.0
mcp==1.15.0
psycopg[binary]==3.2.10
google-auth==2.49.1
